refactor(simulation): structure dashboard application contracts
