+++ src/main/java/com/renewsim/backend/simulation_service/dashboard/application/projection/PortfolioDashboardDistributionByStatus.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.application.dashboard;
+package com.renewsim.backend.simulation_service.dashboard.application.projection;
 
 public record PortfolioDashboardDistributionByStatus(
         String label,
