+++ src/test/java/com/renewsim/backend/simulation_service/create/application/CreateSimulationServiceTest.java
@@ -6,16 +6,17 @@ import com.renewsim.backend.simulation_service.create.application.technology.hyd
 import com.renewsim.backend.simulation_service.create.application.technology.solar.SolarSimulationAssessmentPolicy;
 import com.renewsim.backend.simulation_service.create.application.technology.solar.SolarSimulationEngine;
 import com.renewsim.backend.simulation_service.create.application.technology.wind.WindSimulationEngine;
-import com.renewsim.backend.simulation_service.application.dashboard.GetPortfolioDashboardService;
-import com.renewsim.backend.simulation_service.application.dashboard.PortfolioDashboardRiskAlert;
-import com.renewsim.backend.simulation_service.application.dashboard.PortfolioDashboardResult;
+import com.renewsim.backend.simulation_service.dashboard.application.GetPortfolioDashboardService;
+import com.renewsim.backend.simulation_service.create.application.command.CreateRealSimulationCommand;
+import com.renewsim.backend.simulation_service.dashboard.application.projection.PortfolioDashboardRiskAlert;
+import com.renewsim.backend.simulation_service.dashboard.application.projection.PortfolioDashboardResult;
 import com.renewsim.backend.simulation_service.detail.application.GetSimulationService;
 import com.renewsim.backend.simulation_service.history.application.ListSimulationsService;
-import com.renewsim.backend.simulation_service.history.application.UserSimulationListResult;
-import com.renewsim.backend.simulation_service.application.port.out.PvgisSolarResourcePort;
-import com.renewsim.backend.simulation_service.application.port.out.SimulationRecordRepositoryPort;
-import com.renewsim.backend.simulation_service.application.port.out.TechnologyLookupPort;
-import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
+import com.renewsim.backend.simulation_service.history.application.result.UserSimulationListResult;
+import com.renewsim.backend.simulation_service.create.application.port.out.PvgisSolarResourcePort;
+import com.renewsim.backend.simulation_service.shared.application.port.out.SimulationRecordRepositoryPort;
+import com.renewsim.backend.simulation_service.shared.application.port.out.TechnologyLookupPort;
+import com.renewsim.backend.simulation_service.shared.application.SimulationDetailsResult;
 import com.renewsim.backend.simulation_service.domain.exception.SimulationNotFoundException;
 import com.renewsim.backend.simulation_service.domain.model.Simulation;
 import com.renewsim.backend.simulation_service.domain.model.SimulationId;
